Updated refactorization to be compliant with ODM
