+++ Document/PostManager.php
@@ -21,12 +21,84 @@ use Doctrine\ODM\MongoDB\DocumentManager;
 class PostManager extends BasePostManager
 {
     /**
-     * Get the DB driver connection.
-     *
-     * @return Connection
+     * {@inheritdoc}
      */
     public function getConnection()
     {
         return $this->om->getConnection();
+    }
+
+    /**
+     * @param $year
+     * @param $month
+     * @param $day
+     * @param $slug
+     *
+     * @return mixed
+     */
+    public function findOneBySlug($year, $month, $day, $slug)
+    {
+        $pdqp = $this->getPublicationDateQueryParts(sprintf('%s-%s-%s', $year, $month, $day), 'day');
+
+        return $this->dm->getRepository($this->class)
+            ->createQueryBuilder()
+            ->field('slug')->equals($slug)
+
+            ->andWhere($pdqp['query'])
+            ->getQuery()
+            ->getSingleResult();
+    }
+
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getPager(array $criteria, $page, $maxPerPage = 10)
+    {
+        $parameters = array();
+        $query = $this->dm->getRepository($this->class)
+            ->createQueryBuilder('p')
+            ->select('p, t')
+            ->leftJoin('p.tags', 't')
+            ->orderby('p.publicationDateStart', 'DESC');
+
+        // enabled
+        $criteria['enabled'] = isset($criteria['enabled']) ? $criteria['enabled'] : true;
+        $query->andWhere('p.enabled = :enabled');
+        $parameters['enabled'] = $criteria['enabled'];
+
+        if (isset($criteria['date'])) {
+            $query->andWhere($criteria['date']['query']);
+            $parameters = array_merge($parameters, $criteria['date']['params']);
+        }
+
+        if (isset($criteria['tag'])) {
+            $query->andWhere('t.slug LIKE :tag and t.enabled = :tag_enabled');
+            $parameters['tag'] = $criteria['tag'];
+            $parameters['tag_enabled'] = true;
+        }
+
+        $query->setParameters($parameters);
+
+        $pager = new Pager();
+        $pager->setQuery(new ProxyQuery($query));
+        $pager->setPage($page);
+        $pager->init();
+
+        return $pager;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getPublicationDateQueryParts($date, $step, $alias = 'p')
+    {
+        return array(
+            'query'  => sprintf('%s.publicationDateStart >= :startDate AND %s.publicationDateStart < :endDate', $alias, $alias),
+            'params' => array(
+                'startDate' => new \DateTime($date),
+                'endDate'   => new \DateTime($date . '+1 ' . $step)
+            )
+        );
     }
 }
